+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## TBD
+
+## Bug fixes
+* Enable fastlane to report custom metadata when using `send_build_to_bugsnag` action.
+  [#75](https://github.com/bugsnag/bugsnag-dsym-upload/pull/75)
+
 ## 2.2.1 (19-01-2022)
 
 ### Bug fixes